Cast to double.

+++ src/ui.c
@@ -1047,8 +1047,7 @@ _extra_view_update(Ui *ui, results_t *results)
         evas_object_size_hint_weight_set(progress, EVAS_HINT_EXPAND, EVAS_HINT_EXPAND);
         elm_progressbar_span_size_set(progress, 1.0);
         elm_progressbar_unit_format_set(progress, "%1.2f%%");
-
-        elm_progressbar_value_set(progress, results->power.percent / 100);
+        elm_progressbar_value_set(progress, (double) results->power.percent / 100);
         evas_object_show(progress);
         elm_object_content_set(frame, progress);
         elm_box_pack_end(box, frame);
